extract logger and wait on result to new class
