+++ tests/acceptance/constants.js
@@ -1,1 +1,3 @@
 module.exports.PORT=9999;
+module.exports.VERTICAL_SEARCH_QUERY_URL=/v2\/accounts\/me\/answers\/vertical\/query/;
+module.exports.UNIVERSAL_SEARCH_QUERY_URL=/v2\/accounts\/me\/answers\/query/;
