+++ tests/acceptance/suites/vertical-full-page-map-with-filters.js
@@ -3,16 +3,17 @@ import SearchBar from '../blocks/searchbar';
 import VerticalResults from '../blocks/verticalresults';
 import ThemeMap from '../blocks/thememap';
 import CollapsibleFilters from '../blocks/collapsiblefilters';
+import SearchRequestLogger from '../searchrequestlogger';
 
 fixture`Vertical Full Page Map with Filters and Clusters`
   .page(`http://localhost:${PORT}/locations_full_page_map_with_filters`)
   .beforeEach(async t => {
-    await VerticalResults.registerLogger(t);
+    await SearchRequestLogger.registerVerticalSearchLogger(t);
   });
 
 test('Clicking on a pin closes the filter view', async t => {
   await SearchBar.submitQuery('virginia');
-  await VerticalResults.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   await CollapsibleFilters.viewFilters();
   await ThemeMap.selectPin();
   const isFilterViewOpen = await CollapsibleFilters.isFilterViewOpen();
@@ -21,7 +22,7 @@ test('Clicking on a pin closes the filter view', async t => {
 
 test('Clicking on a cluster causes the map to zoom in', async t => {
   await SearchBar.submitQuery('virginia');
-  await VerticalResults.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   const zoom = await ThemeMap.getZoom();
   await ThemeMap.selectPinCluster();
   const zoomAfterSelectingCluster = await ThemeMap.getZoom();
@@ -30,7 +31,7 @@ test('Clicking on a cluster causes the map to zoom in', async t => {
 
 test('Clicking on a cluster causes a new search to be ran', async t => {
   await SearchBar.submitQuery('virginia');
-  await VerticalResults.waitOnSearchComplete();
+  await SearchRequestLogger.waitOnSearchComplete();
   const numResults = await VerticalResults.getNumResults();
   await ThemeMap.selectPinCluster();
   const numResultsAfterSelectingCluster = await VerticalResults.getNumResults();
